requirements.txt
streamlit
pandas
groq
google-api-python-client
google-auth-httplib2
google-auth-oauthlib